LayoutTests/fast/scrolling/programmatic-smooth-scroll-from-nonzero-offset-expected.txt

PASS observedScrollLeftBelow1000 is false
PASS scrollContainer.scrollLeft is 2000
PASS successfullyParsed is true

TEST COMPLETE
